requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.8.0
numpy>=1.26.0
google-genai>=1.0.0
langchain-text-splitters>=0.3.0
